app/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.1.5
djangorestframework==3.15.2
filelock==3.16.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
load-dotenv==0.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.27.1
